Routes prevention to break un first migration

+++ app/Http/routes.php
@@ -23,20 +23,24 @@ Route::get('/settings/disable-module/{package}', 'SettingsController@disableModu
 
 use App\Http\Helpers\RouteHelper;
 
+//Check if the routes table exist's before using it (For example to protect app to break in the 1º migration stage)
+
+if (Schema::hasTable('routes')) {
 //Inject the routes
-foreach (RouteHelper::fetchRoutes() as $route) {
-    switch ($route->type) {
-        case "GET":
-            Route::get($route->url, $route->action);
-            break;
-        case "POST":
-            Route::post($route->url, $route->action);
-            break;
-        case "PUT":
-            Route::put($route->url, $route->action);
-            break;
-        case "DELETE":
-            Route::delete($route->url, $route->action);
-            break;
+    foreach (RouteHelper::fetchRoutes() as $route) {
+        switch ($route->type) {
+            case "GET":
+                Route::get($route->url, $route->action);
+                break;
+            case "POST":
+                Route::post($route->url, $route->action);
+                break;
+            case "PUT":
+                Route::put($route->url, $route->action);
+                break;
+            case "DELETE":
+                Route::delete($route->url, $route->action);
+                break;
+        }
     }
 }
